+++ app/inscriptions/create/page.tsx
@@ -1,0 +1,8 @@
+import { NextPage } from 'next';
+import { InscriptionsCreate } from '../components/inscription-create';
+
+const InscriptionsCreatePage: NextPage = () => {
+  return <InscriptionsCreate />;
+};
+
+export default InscriptionsCreatePage;